requirements.txt
discord.py
python-dotenv
Pillow
openai
